Changed input method.
The puzzle size is hard-coded to be 9x9 in the constructors, but the
underlying code is agnostic if this so that it will be easy to go back
to accepting puzzles with a different size.

If a file name is given, then that will be read. If no name is given,
then input will be read from stdin.

There is really no error checking. If a file has an error, then the
program will hang.
for other sizes