Modified input format.
Since we are parsing 9x9 puzzles by default, we assume all number
entries are a single digit. This means there does not need to be a
space between numbers. Each digit is considered to be a complete
number. Modified example input files to use the new format. The old
format with spaces is still accepted as can be seen in the Easy
example.